Refactor: split Chrome and Google Docs code into separate modules
- Move Chrome tab interaction to chrome.py
- Move Google Docs integration to gdocs.py
- Alphabetize dependencies in pyproject.toml
