+++ pyproject.toml
@@ -17,10 +17,10 @@ classifiers = [
 ]
 
 dependencies = [
-    "questionary",
     "google-api-python-client",
     "google-auth-httplib2",
     "google-auth-oauthlib",
+    "questionary",
 ]
 
 [project.scripts]
